requirements.txt
python-telegram-bot==13.11
python-dotenv~=0.19.2
requests~=2.27.1